changed slider library, and simplified thumbPicker logic

+++ components/video-thumb-picker.tsx
@@ -1,18 +1,18 @@
 import * as FileSystem from "expo-file-system";
 import { VideoView } from "expo-video";
 import * as VideoThumbnails from "expo-video-thumbnails";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { View } from "react-native";
 
+import { Slider } from "@miblanchard/react-native-slider";
 import { createId } from "@paralleldrive/cuid2";
-import Slider from "@react-native-community/slider";
 import { GestureDetector } from "react-native-gesture-handler";
 import { toast } from "sonner-native";
 
 import { VideoMeta } from "@/db/schema";
 import { useVideoPlayerControls } from "@/hooks/useVideoPlayerControls";
 import { VIDEOS_DIR } from "@/lib/constants";
-import { BanIcon, ImageIcon, SaveIcon } from "@/lib/icons";
+import { ImageDownIcon, LockIcon, LockOpenIcon } from "@/lib/icons";
 import { useVideoStore } from "@/lib/store";
 
 import { Button } from "@/components/ui/button";
@@ -23,7 +23,7 @@ interface VideoThumbPickerProps {
 }
 
 export default function VideoThumbPicker({ videoInfo }: VideoThumbPickerProps) {
-  const [tmpThumbUri, setTmpThumbUri] = useState("");
+  const [isLocked, setIsLocked] = useState(true);
   const [isSaving, setIsSaving] = useState(false);
 
   const updateVideo = useVideoStore((state) => state.updateVideo);
@@ -40,46 +40,23 @@ export default function VideoThumbPicker({ videoInfo }: VideoThumbPickerProps) {
     tapGesture,
   } = useVideoPlayerControls([videoInfo.videoUri], true);
 
-  async function handleGenerateThumb() {
+  async function handleSaveThumb() {
+    if (isSaving || isLocked) return;
+    setIsSaving(true);
+
     try {
       const { uri } = await VideoThumbnails.getThumbnailAsync(videoInfo.videoUri, {
         time: player.currentTime * 1000,
       });
 
-      setTmpThumbUri(uri);
-    } catch (err) {
-      console.error(err);
-      toast.error("Failed to generate thumbnail.");
-    }
-  }
-
-  async function handleCancelThumb() {
-    try {
-      if (tmpThumbUri) {
-        await FileSystem.deleteAsync(tmpThumbUri, { idempotent: true });
-      }
-
-      setTmpThumbUri("");
-    } catch (err) {
-      console.error(err);
-      toast.error("Failed to reset thumbnail.");
-    }
-  }
-
-  async function handleSaveThumb() {
-    if (isSaving) return;
-    setIsSaving(true);
-
-    try {
       const fileId = createId();
       const newUri = `${VIDEOS_DIR}${videoInfo.title}-${fileId}.jpg`;
 
-      await FileSystem.moveAsync({ from: tmpThumbUri, to: newUri });
+      await FileSystem.moveAsync({ from: uri, to: newUri });
       await FileSystem.deleteAsync(videoInfo.thumbUri, { idempotent: true });
 
       await updateVideo({ id: videoInfo.id, values: { thumbUri: newUri } });
-
-      setTmpThumbUri("");
+      toast.success("Thumbnail updated.");
     } catch (err) {
       console.error(err);
       toast.error("Failed to update thumbnail.");
@@ -87,14 +64,6 @@ export default function VideoThumbPicker({ videoInfo }: VideoThumbPickerProps) {
       setIsSaving(false);
     }
   }
-
-  useEffect(() => {
-    return () => {
-      if (tmpThumbUri) {
-        FileSystem.deleteAsync(tmpThumbUri, { idempotent: true }).catch(console.error);
-      }
-    };
-  }, [tmpThumbUri]);
 
   return (
     <GestureDetector gesture={tapGesture}>
@@ -114,62 +83,56 @@ export default function VideoThumbPicker({ videoInfo }: VideoThumbPickerProps) {
             <View className="px-4">
               <Text className="text-sm text-white/70">{time}</Text>
             </View>
-            <View className="flex-1 pb-5">
+            <View className="flex-1 px-4 pb-5">
               <Slider
-                className="h-10 w-full bg-secondary"
+                disabled={isLocked}
                 value={progress}
                 minimumValue={0}
                 maximumValue={1}
                 step={0.01}
+                thumbStyle={{ backgroundColor: isLocked ? "#343434" : "#14b8a6" }}
                 minimumTrackTintColor="#f8fafc"
                 maximumTrackTintColor="#1f242b"
-                tapToSeek
-                onValueChange={onSliderChange}
+                onValueChange={(val) => onSliderChange(Number(val))}
                 onSlidingStart={onSlidingStart}
-                onSlidingComplete={onSlidingComplete}
+                onSlidingComplete={(val) => onSlidingComplete(Number(val))}
               />
             </View>
           </View>
 
           <View className="mb-1 px-4">
-            {tmpThumbUri ? (
-              <View className="flex-row items-center justify-center gap-4">
-                <Button
-                  variant="secondary"
-                  className="flex flex-1 flex-row items-center justify-center gap-4"
-                  onPress={handleCancelThumb}>
-                  <BanIcon
+            <View className="flex-row items-center justify-center gap-4">
+              <Button
+                variant="secondary"
+                className="flex flex-1 flex-row items-center justify-center gap-4"
+                onPress={() => setIsLocked((prev) => !prev)}>
+                {isLocked ? (
+                  <LockIcon
                     className="text-foreground"
                     size={20}
-                    strokeWidth={1.25}
+                    strokeWidth={1.5}
                   />
-                  <Text>Reset</Text>
-                </Button>
-                <Button
-                  className="flex flex-1 flex-row items-center justify-center gap-4"
-                  disabled={isSaving}
-                  onPress={handleSaveThumb}>
-                  <SaveIcon
-                    className="text-background"
+                ) : (
+                  <LockOpenIcon
+                    className="text-foreground"
                     size={20}
-                    strokeWidth={1.25}
+                    strokeWidth={1.5}
                   />
-                  <Text>Save</Text>
-                </Button>
-              </View>
-            ) : (
-              <Button
-                size="lg"
-                className="flex flex-row items-center justify-center gap-4"
-                onPress={handleGenerateThumb}>
-                <ImageIcon
-                  className="text-background"
-                  size={24}
-                  strokeWidth={1.75}
-                />
-                <Text>Generate</Text>
+                )}
+                <Text>{isLocked ? "Unlock" : "Lock"}</Text>
               </Button>
-            )}
+              <Button
+                className="flex flex-1 flex-row items-center justify-center gap-4"
+                disabled={isSaving || isLocked}
+                onPress={handleSaveThumb}>
+                <ImageDownIcon
+                  className="text-background"
+                  size={20}
+                  strokeWidth={1.25}
+                />
+                <Text>Update</Text>
+              </Button>
+            </View>
           </View>
         </View>
       </View>
